Uploaded EXE Version
This version was created using PyInstaller, which means it also can be done yourself if not wanting to trust a random exe, which fair enough. If you want it to have the same results as I did and have the same exe type, you will need to be 'CD'ed into the folder that contains the .py file using CMD (make sure that folder is empty other than that, it creates folders and files in that folder and sub folders), and then use pyinstaller with the file while making sure to add '--onefile -w'. This allows your EXE to become a singular file that can be used anywhere, as well as removing that black window that typically opens with a Python Script/code. Once it finishes, you should be able to use the EXE from wherever. Plus, the exe is smaller than the actual python script file for whatever reason, so that is also a plus, however the difference is only a couple of KB, so not that large of a difference.