A couple minor fixes/changes
Removed some unnecessary lines of code, along with removing the entire creation of both styles to a degree since they weren't even really being used. They are now just place holders of the styles and everything is done using theme_switch. Along with making it so that the true/false buttons display true/false depending on if they are true or false (idea credit to a friend of mine). If anything else was done, it will be in the specific line comments through github explaining the change.