Fixed Cursor Change
This only happened when hovering over a checkbox, but was unintentionally added to github code due to me being in the process of testing it. Now it has the proper interact hand when hovering over any buttons. Whether its open, save, dark mode, search, or true/false buttons. It will display a proper hand interaction cursor.